+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release Notes for Blitz
 
-## 1.7.2 - Unreleased
+## 1.8.0 - Unreleased
+### Added 
+- Added a setting to enable multiple concurrent requests when warming the cache
+- Added a config setting to disable element caches and element query caches from being stored in the database
+
 ### Changed
-- Changed messages and progress bar behaviour in warm cache console command
+- Improved messages and progress bar behaviour in warm cache console command
 
 ## 1.7.1 - 2018-11-20
 ### Fixed